+++ app/src/main/java/com/app/appvenda/modelos/MItemSeletor.java
@@ -6,15 +6,15 @@ package com.app.appvenda.modelos;
 
 public class MItemSeletor {
 
-    private int id;
+    private Long id;
     private String descricao;
     private String codigo;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
